src/test/java/com/sbrf/reboot/atm/cassettes/CassetteTest.java
package com.sbrf.reboot.atm.cassettes;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.List;

import static org.junit.jupiter.api.Assertions.assertThrows;

class CassetteTest {

    class OneHundred extends Banknote {
    }

    class OneThousand extends Banknote {
    }

    @Test
    void getCountBanknotes() {
        OneHundred oneHundred = new OneHundred();

        Cassette<OneHundred> cassette = new Cassette<>(new ArrayList<OneHundred>() {{
            add(oneHundred);
            add(oneHundred);
            add(oneHundred);
//            add(new OneThousand()); //it will not compile
//            add(new Banknote()); //it will not compile
        }});

        Assertions.assertEquals(3, cassette.getCountBanknotes());
    }


    @Test
    void printList() {
        List<Object> list = Arrays.asList("Text1", "Text2");
        for (Object el : list) {
            Assertions.assertEquals(el, Cassette.getValue(el, String.class));
        }
    }
}